+++ api/notation/notation.py
@@ -1,101 +1,16 @@
 from __future__ import annotations
+from api.notation.referentiel import Referentiel
 
 from enum import Enum, unique
 from typing import Tuple, Dict, List
 
-from api.data.generated.referentiels import actions
-from api.models.generated.action_referentiel import ActionReferentiel
 from api.models.generated.action_referentiel_score import ActionReferentielScore
-
-
-class Referentiel:
-    """Referentiel.
-
-    Takes the root action of a referentiel and compute tables and indices.
-    - points are used by Notation to compute scores.
-    - percentages are used to display/test computation
-    - actions allows for index to
-
-    ## Indices
-    indices, forward and backward are indices list for iterating either
-    in no particular order, forward (from root to tâches) or backward (tâches to root) respectively.
-    """
-
-    def __init__(self, root_action: ActionReferentiel):
-        self.root_action: ActionReferentiel = root_action
-        self.points: Dict[Tuple, float] = {}
-        self.percentages: Dict[Tuple, float] = {}
-        self.actions: Dict[Tuple, ActionReferentiel] = {}
-        self.indices: List[Tuple]
-        self.forward: List[Tuple]
-        self.backward: List[Tuple]
-        self.__build_indices()
-        self.__build_points()
-        self.__build_percentages()
-
-    def children(self, parent: tuple) -> List[tuple]:
-        """Returns the children indices"""
-        return [index for index in self.indices if index[:-1] == parent and len(index)]
-
-    def siblings(self, index: tuple) -> List[tuple]:
-        """Returns the index siblings including index"""
-        parent = index[:-1]
-        return self.children(parent)
-
-    def __build_indices(self):
-        """Build all indices lists"""
-        self.indices: List[Tuple] = []
-
-        def add_action(action: ActionReferentiel):
-            index = tuple([element for element in action.id_nomenclature.split('.') if element])
-            self.indices.append(index)
-            self.actions[index] = action
-
-            for action in action.actions:
-                add_action(action)
-
-        add_action(self.root_action)
-
-        self.forward: List[Tuple] = sorted(self.indices, key=lambda i: len(i))
-        self.backward: List[Tuple] = sorted(self.indices, key=lambda i: len(i), reverse=True)
-
-    def __build_points(self):
-        """Build points
-        A référentiel is worth 500 points thus if every actions had a been done
-        perfectly a collectivité would obtain 500 points.
-
-        Axes and orientations points are hardcoded for now as the markdown is
-        not properly defined for now.
-        """
-        for index in self.indices:
-            if len(index) == 0:
-                # référentiel
-                points = 500
-            elif len(index) == 1:
-                # axe
-                points = 100
-            else:
-                # orientation, niveau, tache
-                points = max(self.actions[index].points, 0) * (self.points[index[:-1]] / 100)
-            self.points[index] = points
-
-    def __build_percentages(self):
-        """Build percentages
-
-        Percentages are relative to parents. If an action had 4 children, each would be .25 that is 25%"""
-        for index in self.points.keys():
-            if len(index) > 0:
-                p = self.points[index]
-                if p == 0:
-                    self.percentages[index] = 0
-                else:
-                    self.percentages[index] = self.points[index] / self.points[index[:-1]]
-        self.percentages[tuple()] = 100
 
 
 @unique
 class Statut(Enum):
     """Represent the statut of an action"""
+
     pas_fait = 0
     fait = 1
     pas_concerne = 2
@@ -107,22 +22,22 @@ class Statut(Enum):
 
         Note there is no 'programmée' as it does not count toward notation.
         """
-        if avancement == 'non_concernee':
+        if avancement == "non_concernee":
             return Statut.pas_concerne
-        elif avancement == 'pas_faite':
+        elif avancement == "pas_faite":
             return Statut.pas_fait
-        elif avancement == 'faite':
+        elif avancement == "faite":
             return Statut.fait
         return Statut.vide
 
     def __str__(self):
         if self == Statut.pas_fait:
-            return 'pas_faite'
+            return "pas_faite"
         elif self == Statut.fait:
-            return 'faite'
+            return "faite"
         elif self == Statut.pas_concerne:
-            return 'non_concerne'
-        return ''
+            return "non_concerne"
+        return ""
 
 
 class Notation:
@@ -142,9 +57,15 @@ class Notation:
 
     def reset(self):
         self.potentiels: Dict[Tuple, float] = self.referentiel.points.copy()
-        self.statuts: Dict[Tuple, Statut] = {index: Statut.vide for index in self.referentiel.indices}
-        self.points: Dict[Tuple, float] = {index: .0 for index in self.referentiel.indices}
-        self.percentages: Dict[Tuple, float] = {index: .0 for index in self.referentiel.indices}
+        self.statuts: Dict[Tuple, Statut] = {
+            index: Statut.vide for index in self.referentiel.indices
+        }
+        self.points: Dict[Tuple, float] = {
+            index: 0.0 for index in self.referentiel.indices
+        }
+        self.percentages: Dict[Tuple, float] = {
+            index: 0.0 for index in self.referentiel.indices
+        }
         self.dirty: bool = False
 
     def set_statut(self, index: Tuple, statut: Statut):
@@ -174,7 +95,7 @@ class Notation:
                 referentiel_points=self.referentiel.points[index],
                 referentiel_percentage=self.referentiel.percentages[index],
             )
-            for index in self.referentiel.indices if len(index)
+            for index in self.referentiel.indices
         ]
 
     def __propagate_statuts(self):
@@ -194,8 +115,9 @@ class Notation:
             """parent status from its children"""
             if chidren_statuts.count(Statut.pas_concerne) == len(chidren_statuts):
                 return Statut.pas_concerne
-            elif chidren_statuts.count(Statut.fait) + chidren_statuts.count(Statut.pas_concerne) \
-                    == len(chidren_statuts):
+            elif chidren_statuts.count(Statut.fait) + chidren_statuts.count(
+                Statut.pas_concerne
+            ) == len(chidren_statuts):
                 return Statut.fait
             return Statut.pas_fait
 
@@ -213,7 +135,9 @@ class Notation:
                 continue
             children = self.referentiel.children(index)
             if children:
-                self.statuts[index] = compute_parent_statut([self.statuts[child] for child in children])
+                self.statuts[index] = compute_parent_statut(
+                    [self.statuts[child] for child in children]
+                )
 
     def __compute_potentiels(self):
         """Redistribute `points` of actions with a `statuts non concerné`
@@ -230,15 +154,20 @@ class Notation:
                 continue
             elif exclusions == len(children):
                 for child in children:
-                    self.potentiels[child] = .0
+                    self.potentiels[child] = 0.0
             else:
-                excluded = sum([self.referentiel.points[child] for child in children if
-                                self.statuts[child] == Statut.pas_concerne])
+                excluded = sum(
+                    [
+                        self.referentiel.points[child]
+                        for child in children
+                        if self.statuts[child] == Statut.pas_concerne
+                    ]
+                )
                 redistribution = excluded / (len(children) - exclusions)
 
                 for child in children:
                     if self.statuts[child] == Statut.pas_concerne:
-                        self.potentiels[child] = .0
+                        self.potentiels[child] = 0.0
                     else:
                         self.potentiels[child] += redistribution
 
@@ -246,7 +175,7 @@ class Notation:
         """Compute points from potentiels the propagate the sums"""
         # first pass
         for index in self.referentiel.indices:
-            progress = 1.0 if self.statuts[index] == Statut.fait else .0
+            progress = 1.0 if self.statuts[index] == Statut.fait else 0.0
             self.points[index] = progress * self.potentiels[index]
 
         # second pass
@@ -260,6 +189,3 @@ class Notation:
         for index in self.referentiel.indices:
             if self.potentiels[index] != 0:
                 self.percentages[index] = self.points[index] / self.potentiels[index]
-
-
-referentiel_eci = Referentiel(next(action for action in actions if action.id.startswith('economie_circulaire')))
